% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cred.R, R/log_query.R
\name{cred}
\alias{cred}
\alias{cred_env}
\alias{cred_file}
\alias{cred_json}
\alias{qlog_loud}
\alias{qlog_silent}
\title{Create databse credentials object}
\usage{
cred(user, password, host, dbname, port = NA, engine = NA, .qlog = NA,
  .show_warn = NA)

cred_json(json)

cred_file(file)

cred_env(var_name)

qlog_loud(.db = NULL)

qlog_silent(.db = NULL)
}
\arguments{
\item{user}{database username}

\item{password}{databse password}

\item{host}{database server hostname}

\item{dbname}{database name to connect to}

\item{port}{databse server port number}

\item{engine}{specify database engine to use}

\item{.qlog}{force logging of queries to console}

\item{.show_warn}{show warnings from running query, usually typecasting.
See \code{\link{dbwarn}}}

\item{json}{JSON credentials string.}

\item{file}{filepath to JSON credentials file}

\item{var_name}{system variable name}

\item{.db}{credentials used for this query, if \code{NULL}, default is used}
}
\value{
a \code{db_credentials} object
}
\description{
Combine all the required elements to create a databse connection into
a single object. Options can be set interatively via \code{\link{qlog_loud}},
\code{\link{qlog_silent}}, and  \code{\link{dbwarn}}.

Retrieve database credentials from a structured JSON and create a
db_credentials object

Retrieve database credentials from a file. Uses cred_json to parse.

Retrieve database credentials from a system environment variable.
Uses cred_json to parse.

Force queries to be logged or not, regardless of the other settings.
}
\examples{
cred("bob","passwd","127.0.0.1","foo")

json <- paste0('{"user": "usr", "password": "pass123", "host": "127.0.0.1",',
               ' "dbname": "foo"}')
foo <- cred_json(json)
}

