% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_cred}
\alias{db_cred}
\alias{db_file}
\alias{db_json}
\title{Register Database}
\usage{
db_cred(cred)

db_json(json)

db_file(file)
}
\arguments{
\item{cred}{credentials object}

\item{json}{a JSON credentials string.}

\item{file}{a file containing a JSON credentials string.}
}
\description{
Register database to create a connection and use the database with \code{\link{gq}}.
}

