% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{db_date}
\alias{db_date}
\title{Format date for database query}
\usage{
db_date(date)
}
\arguments{
\item{date}{date to be formatted.}
}
\value{
date formatted as a string for database.
}
\description{
Format a date to use in a databse query. Includes quotes.
}
\examples{
db_date(lubridate::now())
\dontrun{
gq("SELECT * FROM table WHERE date > ", db_date(lubridate::now()))
}

}

