% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{trans_begin}
\alias{trans_begin}
\alias{trans_commit}
\alias{trans_rollback}
\title{Begin/commit/rollback SQL transactions}
\usage{
trans_begin(.db = NULL)

trans_commit(force = FALSE, .db = NULL)

trans_rollback(force = FALSE, .db = NULL)
}
\arguments{
\item{.db}{the connection to use throughout the transaction}

\item{force}{force the action to take place}
}
\description{
db_trans require an open connection, so you will need to manually create the connection,
preform queries with \code{\link{gq}} and then and close the
connection.

Commiting an SQL transaction saves the actions made durreing the transaction to the database.

Rollback resets any db_trans that might have happened durring the transaction.
}

