% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gq.R
\name{dbwarn}
\alias{dbwarn}
\title{Force database warnings to be shown}
\usage{
dbwarn(.db = NULL)
}
\arguments{
\item{.db}{credentials used for this query, if \code{NULL}, default is used}
}
\description{
Force database warnings to be shown. Usually they are typecasting warnings,
and so are suppressed by default. This is a good debugging tool if queries
are throwing errors or not returning the expected output.
}

