% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gq.R
\name{gq}
\alias{gq}
\title{Get query from database}
\usage{
gq(..., .db = NULL)
}
\arguments{
\item{...}{strings to be combined to form SQL query. Query is combined with \code{paste0(..., collapse = " ")}.}

\item{.db}{database credentials to be used for the query.}
}
\value{
a \code{data.frame} with the results of the query.
}
\description{
Retrieve query from registered database, or supply alternate credentials or connection
(\code{.cred} or \code{.conn} respectively.)
}

