% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{list_dbs}
\alias{list_dbs}
\title{Get List of Database Credentials}
\usage{
list_dbs()
}
\description{
Get a list of databases registered with the package. All of these databases have open
connections.
}
\examples{
list_dbs()

}

