% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_query.R
\name{log_query}
\alias{log_query}
\title{Log the any query to the console}
\usage{
log_query(query, .db)
}
\arguments{
\item{query}{the query string}

\item{.db}{The database the query is being sent to}
}
\value{
None
}
\description{
Print a query to the console, along with time of query and the connection
details of the database connection used.
}

