% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{parse_date}
\alias{parse_date}
\title{Convert database date string to an R date}
\usage{
parse_date(date)
}
\arguments{
\item{date}{date string passed to \code{ymd}}
}
\description{
Convert database date format to an R date format. Timezone comes from
\code{dbezr.timezone} option.
}

