% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{parse_date_time}
\alias{parse_date_time}
\title{Convert database DateTime string to an R date}
\usage{
parse_date_time(date_time)
}
\arguments{
\item{date_time}{date_time string passed to \code{parse_date_time}}
}
\description{
Convert database DateTime format to an R date format. Timezone comes from
\code{dbezr.timezone} option.
}

