% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\alias{prep.POSIXct}
\alias{prep.POSIXlt}
\alias{prep.character}
\alias{prep.default}
\alias{prep.logical}
\alias{prep.numeric}
\alias{prep_nulls}
\title{Prepare data for Entry into database}
\usage{
prep(x, .db)

\method{prep}{character}(x, .db = NULL)

\method{prep}{POSIXct}(x, .db)

\method{prep}{POSIXlt}(x, .db)

\method{prep}{numeric}(x, .db)

\method{prep}{logical}(x, .db)

\method{prep}{default}(x, .db)

prep_nulls(x, orig)
}
\arguments{
\item{x}{vector to prepare for database}

\item{.db}{the connection to use throughout the transaction}

\item{orig}{original values}
}
\description{
Prepare a vector of data to be inserted into the database
}

