% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cred.R
\name{print.db_credentials}
\alias{print.db_credentials}
\title{Print db_credentials object}
\usage{
\method{print}{db_credentials}(x, show_password = FALSE, ...)
}
\arguments{
\item{x}{db_credentials to print.}

\item{show_password}{show password, defaults to \code{FALSE}.}

\item{...}{ignored.}
}
\description{
Print a db_credentials object, traditionally hiding the password.
}

