% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{print.db_credentials_list}
\alias{print.db_credentials_list}
\title{Print list of database credentials}
\usage{
\method{print}{db_credentials_list}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Formats a list of database credentials for printing nicely by printing the individual databases
without the list structure.
}
\examples{
list_dbs()

}

