% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rm_db}
\alias{rm_db}
\alias{rm_db_all}
\alias{rm_db_id}
\title{Remove db object}
\usage{
rm_db(cred)

rm_db_id(id)

rm_db_all()
}
\arguments{
\item{cred}{credentials object}

\item{id}{credential id string generated by cred_id}
}
\value{
\code{NULL}
}
\description{
Removes a db from the registered list, disconnecting it's connection.
}
\examples{
\dontrun{
rm_db(cred)
}

}

