% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{shw_cols}
\alias{shw_cols}
\title{Show columns from a table in the databse}
\usage{
shw_cols(table, schema = NA, .db = NULL)
}
\arguments{
\item{table}{table name to query for}

\item{schema}{schema to show tables from}

\item{.db}{credentials to be used by \code{gq}}
}
\description{
Show a list of tables in the databse
}

