% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{shw_tbls}
\alias{shw_tbls}
\title{Show tables in databse}
\usage{
shw_tbls(schema = NA, .db = NULL)
}
\arguments{
\item{schema}{schema to show tables from}

\item{.db}{credentials to be used by \code{gq}}
}
\description{
Show a list of tables in the databse
}

