% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cred.R
\name{cred}
\alias{cred}
\alias{cred_json}
\title{Create databse credentials object}
\usage{
cred(user, password, host, dbname, port = NA, engine = NA, .qlog = NA,
  .show_warn = NA)

cred_json(json)
}
\arguments{
\item{user}{database username}

\item{password}{databse password}

\item{host}{database server hostname}

\item{dbname}{database name to connect to}

\item{port}{databse server port number}

\item{engine}{specify database engine to use}

\item{.qlog}{force logging of queries to console}

\item{.show_warn}{show warnings from running query, usually typecasting.
See \code{\link{dbwarn}}}

\item{json}{JSON credentials string.}
}
\value{
a \code{db_credentials} obeject
}
\description{
Combine all the required elements to create a databse connection into
a single object. Options can be set interatively via \code{\link{qlog_loud}},
\code{\link{qlog_silent}}, and  \code{\link{dbwarn}}.

Retrive database credentials from a structured JSON and create a
db_credentials object
}
\examples{
cred("bob","passwd","127.0.0.1",3306,"foo")

tmp <- cred_json('{"user": "usr", "password": "pass123", "host": "127.0.0.1", "dbname": "foo"}') # nolint
}

